organized some code
offloaded player movement code to TransmutationBase, which will also be the class that changes when we transmute
changed some names around, like input handler to receiver since it literally does not handle input
added defaults to IInteractable, also the name has an I in from of it now cuz it's an interface
changed PlayerBase to Player since Base implies that it's supposed to be inherited, when that's not the intention
