+++ Assets/Players/TransmutationBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f846fccd00d9cf34baa1b0f2b08187ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
