+++ Assets/Props/IInteractable.cs.meta
