--- Assets/Props/Interactable.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IInteractable
-{
-    public string InteractionPrompt { get; }
-    public Sprite InteractionSprite { get; }
-
-    public void Selected();
-
-    public bool Interact();
-}
